Factor out common value-based identifier creation
