+++ src/Identifier/ElementParameterIdentifierFactory.php
@@ -2,21 +2,14 @@
 
 namespace webignition\BasilModelFactory\Identifier;
 
-use webignition\BasilModel\Identifier\Identifier;
 use webignition\BasilModel\Identifier\IdentifierInterface;
 use webignition\BasilModel\Identifier\IdentifierTypes;
 use webignition\BasilModelFactory\IdentifierTypeFinder;
 use webignition\BasilModelFactory\ValueFactory;
 
-class ElementParameterIdentifierFactory implements IdentifierTypeFactoryInterface
+class ElementParameterIdentifierFactory extends AbstractValueBasedIdentifierFactory implements
+    IdentifierTypeFactoryInterface
 {
-    private $valueFactory;
-
-    public function __construct(ValueFactory $valueFactory)
-    {
-        $this->valueFactory = $valueFactory;
-    }
-
     public static function createFactory()
     {
         return new ElementParameterIdentifierFactory(
@@ -47,15 +40,6 @@ class ElementParameterIdentifierFactory implements IdentifierTypeFactoryInterfac
 
         $identifierString = trim($identifierString);
 
-        $identifier = new Identifier(
-            IdentifierTypes::ELEMENT_PARAMETER,
-            $this->valueFactory->createFromValueString($identifierString)
-        );
-
-        if (null !== $name) {
-            $identifier = $identifier->withName($name);
-        }
-
-        return $identifier;
+        return $this->createForType($identifierString, IdentifierTypes::ELEMENT_PARAMETER, $name);
     }
 }
